requirements.txt
streamlit
torch
torchvision
pillow
matplotlib
fpdf
pandas